requirements.txt
Flask
pydantic
pydantic_core
pydantic_ai
PyMuPDF
nest_asyncio
gunicorn
python-dotenv
